requirements.txt
# python==3.10.4
numpy==1.21.5
matplotlib==3.5.2
rospkg==1.4.0
scipy==1.7.3
ruamel.yaml==0.16.12